sama/requirements.txt
openai-whisper
psycopg2-binary
paho-mqtt
sounddevice 
scipy